Changed all code to use f32 instead of f64 for significant performance gains

+++ src/activation/mod.rs
@@ -7,8 +7,8 @@ pub mod sigmoid;
 pub mod tanh;
 
 pub trait ActivationFunction {
-    fn func(x: f64) -> f64;
-    fn deriv(x: f64) -> f64;
+    fn func(x: f32) -> f32;
+    fn deriv(x: f32) -> f32;
 }
 
 pub fn func_all<
@@ -16,8 +16,8 @@ pub fn func_all<
     const C: usize,
     F: ActivationFunction,
 >(
-    x: &SMatrix<f64, R, C>,
-) -> SMatrix<f64, R, C> {
+    x: &SMatrix<f32, R, C>,
+) -> SMatrix<f32, R, C> {
     let mut x = x.clone();
     x.iter_mut().for_each(|xi| *xi = F::func(*xi));
     x
@@ -28,8 +28,8 @@ pub fn deriv_all<
     const C: usize,
     F: ActivationFunction,
 >(
-    x: &SMatrix<f64, R, C>,
-) -> SMatrix<f64, R, C> {
+    x: &SMatrix<f32, R, C>,
+) -> SMatrix<f32, R, C> {
     let mut x = x.clone();
     x.iter_mut().for_each(|xi| *xi = F::deriv(*xi));
     x
