+++ src/activation/tanh.rs
@@ -6,11 +6,11 @@ use super::ActivationFunction;
 pub struct Tanh;
 
 impl ActivationFunction for Tanh {
-    fn func(x: f64) -> f64 {
+    fn func(x: f32) -> f32 {
         x.tanh()
     }
 
-    fn deriv(x: f64) -> f64 {
+    fn deriv(x: f32) -> f32 {
         1. - x.tanh().powi(2)
     }
 }
