+++ src/layers/embedding.rs
@@ -27,24 +27,24 @@ impl<const N: usize> Embedding<N> {
     pub fn embed<const T: usize>(
         &self,
         words: &str,
-    ) -> [SVector<f64, T>; N] {
+    ) -> [SVector<f32, T>; N] {
         let words = words.split(' ').map(|word| match self
             .embeddings
             .embedding(word)
         {
             Some(x) => {
-                SVector::<f64, T>::from_column_slice(
+                SVector::<f32, T>::from_column_slice(
                     &x.as_slice()
                         .unwrap_or(&[0.])
                         .into_iter()
-                        .map(|&x| x as f64)
+                        .map(|&x| x as f32)
                         .collect::<Vec<_>>(),
                 )
             }
-            None => SVector::<f64, T>::zeros(),
+            None => SVector::<f32, T>::zeros(),
         });
 
-        let mut out = [SVector::<f64, T>::zeros(); N];
+        let mut out = [SVector::<f32, T>::zeros(); N];
         words.take(N).enumerate().for_each(|(i, x)| {
             out[i] = x;
         });
