+++ src/loss/crossent.rs
@@ -6,18 +6,18 @@ pub struct CrossEntropy;
 
 impl<const N: usize> LossFunction<N> for CrossEntropy {
     fn func(
-        y_out: SVector<f64, N>,
-        y_test: SVector<f64, N>,
-    ) -> f64 {
+        y_out: SVector<f32, N>,
+        y_test: SVector<f32, N>,
+    ) -> f32 {
         -(0..N)
             .map(|i| y_test[i] * y_out[i].log2())
-            .sum::<f64>()
+            .sum::<f32>()
     }
 
     fn grad(
-        y_out: SVector<f64, N>,
-        y_test: SVector<f64, N>,
-    ) -> SVector<f64, N> {
+        y_out: SVector<f32, N>,
+        y_test: SVector<f32, N>,
+    ) -> SVector<f32, N> {
         let out = (0..N).map(|i| -y_test[i] / y_out[i]);
         SVector::from_iterator(out)
     }
